archive/pyproject.toml
[project]
name = "structured-data-jobs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.14",
    "alembic>=1.15.1",
    "asyncpg>=0.30.0",
    "atlas-provider-sqlalchemy>=0.2.4",
    "azure-identity>=1.21.0",
    "black>=25.1.0",
    "boto2>=0.1.0",
    "boto3>=1.37.23",
    "flake8>=7.1.2",
    "html2text>=2024.2.26",
    "ipdb>=0.13.13",
    "isort>=6.0.1",
    "logfire[asyncpg,sqlalchemy]>=3.8.1",
    "openai>=1.66.3",
    "playwright>=1.50.0",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.39",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.1.2",
]
